Site landing content (#39)
* Add new filesystem source

* Add collection mdx files

* Add to query to pull collections data

The mdx and jpg files in collections contain the information for the
landing page of each section of the site. Added a bit more graphql to
the query as well as one new piece of frontmatter to get this info.
Displaying it on the page was easy, awesome, and it just worked!