Add Pagination for collections and posts (#45)
* Update excerpt so it isn't duplicate

* Just pass excerpt from node

It will be the frontmatter.excerpt if it is provided!

* Create Pagination component

Steal is probably a better word. This is from Skillthrive.
Modified a bit to use my colors.

* Add Pagination to post list page

Bring in calculations and modify to use collection.
Only display collection intro on first page.

* Tweaking date so it sorts clearly

* Add next and previous info for a post

* Add pagination for posts

This is so different than pages. Pagination is doing too much.

For now, it is working so if I accidentally make more posts,
it'll be okay!