Tweak ssr and hero post (#66)
* Add Mdx components to SSR

Accidentally ran a build and saw warnings about not having
the Mdx components in the global space.
The components object in gatsby-browser.js had them but
not in gatsby-ssr.js. This is remedied.

* Remove the '- Part' and just put I

It looked too busy with Blah Blah - Part I
Isn't Blah Blah I much better?
It is confusing but only until part 2 shows up!