Fix big Logo issue (#78)
Learned that SVG has its own thoughts about how big something should be.
It was recommended to wrap the SVG in a div that specifies the proper
width and height. I did that and it seemed to help.
I still see the logo evolving as fonts are applied but alas, for now.