Front page tags (#163)
* Add missing tags

* Add list of tags to front page

They don't link to anything but they each indicate how many pages they
tag.

* Remove import of prop-type array

`array` is not used in this component. Nor are prop-types!

* Add tag pages

I did too much in this commit but I can't see how to make it clear.
Lesson should be learned! Plus it is late and I should be asleep!

- Extracted tagData function for reuse
- Used module.exports so I can use in `gatsby-node`
- Create tag pages
- Create tag template that simply uses Posts style

* Add pages for tags

- single-tag template
- Display PostList for each tag
- A PostListCard for each post
- TagLinks for each PostListCard
- TagLink for each tag in a post

* Add collection to card

- Display at an angle for fun on larger screens.
- Overlay the cover picture on very small screens.

* Add tags to posts

- Remove background from TagLinks component.
- Make tag text `xs`